prototype/requirements.txt
flask>=3.0.0
python-nmap>=0.7.1
gunicorn>=21.2.0
